requirements.txt
ephem
PySide6
urllib3
numpy
qt-material
pyserial
matplotlib
cartopy
pyproj
Flask
Flask-SocketIO
requests
python-socketio[client]
sounddevice
pynmea2
